# openapi2-run.yaml
...
swagger: "2.0"
info:
  title: "Coway Meeting Notes & Diarization Agent Gateway"
  description: "Unified Enterprise Ingress Gateway for Coway AI Meeting Notes, Speaker Diarization, and Template Management."
  version: "1.0.0"
schemes:
  - "https"
produces:
  - "application/json"
consumes:
  - "application/json"
paths:
...
  /api/v1/notes/current:
    get:
      summary: "현재 회의록 및 전사본 전문 조회"
      operationId: "getCurrentNotes"
      x-google-backend:
        address: "${BACKEND_CLOUD_RUN_URL}/api/v1/notes/current"
        jwt_audience: "${BACKEND_CLOUD_RUN_URL}"
      parameters:
        - name: "sample"
          in: "query"
          required: false
          type: "string"
      responses:
        200:
          description: "성공"

  # ============================================================================
  # 2. Media Processing & AI Notes Generation Endpoints (GCS Resumable Direct Upload)
  # ============================================================================
...